requirements.txt
streamlit>=1.22
pymupdf>=1.22.0
sentence-transformers>=2.2.2
faiss-cpu>=1.7.4
transformers>=4.34.0
torch>=1.12.0
numpy>=1.24
python-dotenv>=0.21.0
requests>=2.28.0
pillow>=9.0.0
reportlab>=3.6.12
